+++ vllm/model_executor/layers/pool/abstract.py
@@ -2,33 +2,28 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 from abc import ABC, abstractmethod
 from collections.abc import Set
-from typing import TypeAlias
 
 import torch
 import torch.nn as nn
 
 from vllm.model_executor.layers.pool.common import PoolingParamsUpdate
 from vllm.tasks import PoolingTask
+from vllm.v1.outputs import PoolerOutput
 from vllm.v1.pool.metadata import PoolingMetadata
 
-TokenPoolingMethodOutput: TypeAlias = torch.Tensor | list[torch.Tensor]
-"""Applicable to pooling strategies that output one token."""
 
-TokenwisePoolingMethodOutput: TypeAlias = list[torch.Tensor] | list[torch.Tensor | None]
-"""Applicable to pooling strategies that output multiple tokens."""
+class Pooler(nn.Module, ABC):
+    """The interface required for all poolers used in pooling models in vLLM."""
 
-TokenwisePoolingMethodOutputItem: TypeAlias = torch.Tensor | None
-"""Represents a single element of `TokenwisePoolingMethodOutput`."""
-
-PoolingMethodOutput: TypeAlias = TokenPoolingMethodOutput | TokenwisePoolingMethodOutput
-
-
-class PoolingMethod(nn.Module, ABC):
     @abstractmethod
     def get_supported_tasks(self) -> Set[PoolingTask]:
+        """Determine which pooling tasks are supported."""
         raise NotImplementedError
 
     def get_pooling_updates(self, task: PoolingTask) -> PoolingParamsUpdate:
+        """
+        Construct the updated pooling parameters to use for a supported task.
+        """
         return PoolingParamsUpdate()
 
     @abstractmethod
@@ -36,5 +31,8 @@ class PoolingMethod(nn.Module, ABC):
         self,
         hidden_states: torch.Tensor,
         pooling_metadata: PoolingMetadata,
-    ) -> PoolingMethodOutput:
+    ) -> PoolerOutput:
         raise NotImplementedError
+
+
+__all__ = ["Pooler"]
