--- vllm/model_executor/layers/pool/heads/__init__.py
@@ -1,26 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-from .base import (
-    PoolerHead,
-    PoolerHeadOutput,
-    TokenPoolerHeadOutput,
-    TokenwisePoolerHeadOutput,
-)
-from .token import EmbeddingPoolerHead, TokenPoolerHead
-from .tokenwise import (
-    TokenClassifierPoolerHead,
-    TokenEmbeddingPoolerHead,
-    TokenwisePoolerHead,
-)
-
-__all__ = [
-    "PoolerHead",
-    "PoolerHeadOutput",
-    "TokenPoolerHeadOutput",
-    "TokenwisePoolerHeadOutput",
-    "TokenPoolerHead",
-    "EmbeddingPoolerHead",
-    "TokenwisePoolerHead",
-    "TokenEmbeddingPoolerHead",
-    "TokenClassifierPoolerHead",
-]
