--- vllm/model_executor/layers/pool/methods/__init__.py
@@ -1,25 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-from .base import (
-    PoolingMethod,
-    PoolingMethodOutput,
-    TokenPoolingMethodOutput,
-    TokenwisePoolingMethodOutput,
-    TokenwisePoolingMethodOutputItem,
-)
-from .factory import get_token_pooling_method
-from .token import CLSPool, LastPool, MeanPool
-from .tokenwise import AllPool
-
-__all__ = [
-    "PoolingMethod",
-    "TokenPoolingMethodOutput",
-    "TokenwisePoolingMethodOutput",
-    "TokenwisePoolingMethodOutputItem",
-    "PoolingMethodOutput",
-    "CLSPool",
-    "LastPool",
-    "MeanPool",
-    "AllPool",
-    "get_token_pooling_method",
-]
