--- vllm/model_executor/layers/pool/poolers/__init__.py
@@ -1,29 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-from .base import Pooler, PoolerOutput, TokenPoolerOutput, TokenwisePoolerOutput
-from .common import DispatchPooler, DummyPooler, SimplePooler
-from .factory import (
-    pooler_for_classify,
-    pooler_for_embed,
-    pooler_for_token_classify,
-    pooler_for_token_embed,
-)
-from .token import ClassifierPooler
-from .tokenwise import AllPooler, StepPooler
-
-__all__ = [
-    "Pooler",
-    "TokenPoolerOutput",
-    "TokenwisePoolerOutput",
-    "PoolerOutput",
-    "DispatchPooler",
-    "SimplePooler",
-    "DummyPooler",
-    "ClassifierPooler",
-    "AllPooler",
-    "StepPooler",
-    "pooler_for_classify",
-    "pooler_for_embed",
-    "pooler_for_token_classify",
-    "pooler_for_token_embed",
-]
