--- vllm/model_executor/layers/pool/poolers/base.py
@@ -1,43 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-from abc import ABC, abstractmethod
-from collections.abc import Set
-from typing import TypeAlias
-
-import torch
-import torch.nn as nn
-
-from vllm.model_executor.layers.pool.common import PoolingParamsUpdate
-from vllm.tasks import PoolingTask
-from vllm.v1.pool.metadata import PoolingMetadata
-
-TokenPoolerOutput: TypeAlias = torch.Tensor | list[torch.Tensor]
-"""Applicable to pooling strategies that output one token."""
-
-TokenwisePoolerOutput: TypeAlias = list[torch.Tensor] | list[torch.Tensor | None]
-"""Applicable to pooling strategies that output multiple tokens."""
-
-PoolerOutput: TypeAlias = TokenPoolerOutput | TokenwisePoolerOutput
-
-
-class Pooler(nn.Module, ABC):
-    """The interface required for all poolers used in pooling models in vLLM."""
-
-    @abstractmethod
-    def get_supported_tasks(self) -> Set[PoolingTask]:
-        """Determine which pooling tasks are supported."""
-        raise NotImplementedError
-
-    def get_pooling_updates(self, task: PoolingTask) -> PoolingParamsUpdate:
-        """
-        Construct the updated pooling parameters to use for a supported task.
-        """
-        return PoolingParamsUpdate()
-
-    @abstractmethod
-    def forward(
-        self,
-        hidden_states: torch.Tensor,
-        pooling_metadata: PoolingMetadata,
-    ) -> PoolerOutput:
-        raise NotImplementedError
