+++ vllm/model_executor/layers/pool/token/poolers.py
@@ -1,29 +1,69 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-from collections.abc import Callable, Set
+from abc import abstractmethod
+from collections.abc import Set
 from typing import TypeAlias
 
 import torch
 
-from vllm.config import get_current_vllm_config
+from vllm.config import PoolerConfig, get_current_vllm_config
 from vllm.model_executor.layers.pool.activations import (
     PoolerActivation,
     resolve_classifier_act_fn,
 )
-from vllm.model_executor.layers.pool.common import ClassifierFn
-from vllm.model_executor.layers.pool.methods import TokenPoolingMethodOutput
+from vllm.model_executor.layers.pool.common import ClassifierFn, PoolingParamsUpdate
 from vllm.tasks import PoolingTask
 from vllm.v1.pool.metadata import PoolingMetadata
 
-from .base import Pooler, TokenPoolerOutput
+from ..abstract import Pooler
+from .heads import EmbeddingPoolerHead, TokenPoolerHead
+from .methods import TokenPoolingMethod, get_token_pooling_method
 
-TokenPoolingMethod: TypeAlias = Callable[
-    [torch.Tensor, PoolingMetadata],
-    TokenPoolingMethodOutput,
-]
+TokenPoolerOutput: TypeAlias = torch.Tensor | list[torch.Tensor]
 
 
-class ClassifierPooler(Pooler):
+class TokenPooler(Pooler):
+    @abstractmethod
+    def forward(
+        self,
+        hidden_states: torch.Tensor,
+        pooling_metadata: PoolingMetadata,
+    ) -> TokenPoolerOutput:
+        raise NotImplementedError
+
+
+class SimplePooler(TokenPooler):
+    """A layer that pools specific information from hidden states.
+
+    This layer does the following:
+    1. Extracts specific tokens or aggregates data based on pooling method.
+    2. Normalizes output if specified.
+    3. Returns structured results as `PoolerOutput`.
+    """
+
+    def __init__(self, pooling: TokenPoolingMethod, head: TokenPoolerHead) -> None:
+        super().__init__()
+
+        self.pooling = pooling
+        self.head = head
+
+    def get_supported_tasks(self) -> Set[PoolingTask]:
+        return self.pooling.get_supported_tasks()
+
+    def get_pooling_updates(self, task: PoolingTask) -> PoolingParamsUpdate:
+        return self.pooling.get_pooling_updates(task)
+
+    def forward(
+        self,
+        hidden_states: torch.Tensor,
+        pooling_metadata: PoolingMetadata,
+    ) -> TokenPoolerOutput:
+        pooled_data = self.pooling(hidden_states, pooling_metadata)
+        pooled_data = self.head(pooled_data, pooling_metadata)
+        return pooled_data
+
+
+class ClassifierPooler(TokenPooler):
     """A pooling layer for classification tasks.
 
     This layer does the following:
@@ -85,3 +125,20 @@ class ClassifierPooler(Pooler):
 
         # scores shape: [batchsize, num_labels]
         return scores
+
+
+def pooler_for_embed(pooler_config: PoolerConfig):
+    pooling = get_token_pooling_method(pooler_config.get_pooling_type())
+    head = EmbeddingPoolerHead()
+
+    return SimplePooler(pooling=pooling, head=head)
+
+
+def pooler_for_classify(
+    pooler_config: PoolerConfig,
+    classifier: ClassifierFn | None,
+    act_fn: PoolerActivation | str | None = None,
+):
+    pooling = get_token_pooling_method(pooler_config.get_pooling_type())
+
+    return ClassifierPooler(pooling=pooling, classifier=classifier, act_fn=act_fn)
