+++ vllm/model_executor/layers/pool/universal/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+from .poolers import DispatchPooler, DummyPooler
+
+__all__ = ["DispatchPooler", "DummyPooler"]
